bugreporter/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
future==0.17.1
idna==2.8
oauthlib==3.0.1
py-trello==0.14.0
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-telegram-bot==11.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tinydb==3.12.2
urllib3==1.24.1
flake8==3.7.5